add float on sks realisasi

+++ backendapp/api/models.py
@@ -79,7 +79,7 @@ class SuratPenugasan(models.Model):
 
 class PenugasanPengajaran(models.Model):
 	created_at = models.DateTimeField(default=timezone.now)
-	sks_realisasi = models.IntegerField(default=0)
+	sks_realisasi = models.FloatField(default=0)
 	surat_penugasan = models.ForeignKey(
 			SuratPenugasan,
 			on_delete=models.CASCADE,
